# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LumAgg API
  description: |
    Stellar DEX aggregator — quotes and unsigned swap transactions across Soroban AMMs
    (Soroswap, Aquarius, Phoenix, Sushi V3, Comet) and optionally Classic SDEX.

    **Rate limits:** 10 req/s per IP; 60 req/s with valid `X-API-Key` (partner keys).
    See [integrator-guide.md](./integrator-guide.md).
  version: 1.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.lumagg.xyz
    description: Production
  - url: http://127.0.0.1:3100
    description: Local dev
...
paths:
...
  /api/v1/stats:
    get:
      tags: [health]
      summary: On-chain aggregator stats (indexer DB)
      description: |
        Daily rollups from `analytics-indexer` SQLite. Requires `INDEXER_DB_PATH` on the API server.
        Optional `day=YYYY-MM-DD` filters to a single UTC day. Set `format=csv` for grant report download.
        Routed volume is the sum of actual inputs processed by successful DEX legs, priced by each leg's input token.
        Round-trip gross surplus is actual on-chain `amount_out - amount_in`; transaction fees are not deducted.
      parameters:
        - name: day
          in: query
          required: false
          schema:
            type: string
            example: "2026-07-13"
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Stats rollup (JSON or CSV)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
            text/csv:
              schema:
                type: string
        '503':
          description: INDEXER_DB_PATH not configured
...
  /api/v1/arbitrage/stats:
    get:
      tags: [health]
      summary: Aggregated arbitrage statistics
      description: |
        Time-bucketed confirmed round-trip results from the analytics indexer.
        Surplus values are raw atomic units for XLM and USDC and only successful
        transactions contribute to surplus. Requires `INDEXER_DB_PATH` on the API server.
      parameters:
        - name: granularity
          in: query
          required: false
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
          description: Bucket size. The default window is selected from the granularity.
        - name: start
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Inclusive Unix timestamp in seconds.
        - name: end
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Exclusive Unix timestamp in seconds. Defaults to the current time.
      responses:
        '200':
          description: Aggregated arbitrage buckets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArbitrageStatsResponse'
        '400':
          description: Invalid granularity or timestamp range
        '503':
          description: INDEXER_DB_PATH not configured
...
components:
...
  schemas:
...
    StatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            db_path:
              type: string
            invocation_count:
              type: integer
            cursor_ledger:
              type: integer
            oldest_created_at:
              type: integer
            daily:
              type: array
              items:
                type: object
                properties:
                  day:
                    type: string
                  tx_count:
                    type: integer
                  unique_users:
                    type: integer
                  total_amount_in:
                    type: string
                  total_amount_in_usd:
                    type: number
                    nullable: true
                  total_routed_dex_volume:
                    type: string
                    description: Sum of actual DEX leg inputs in mixed atomic units; use USD or by_token for reporting
                  total_routed_dex_volume_usd:
                    type: number
                    nullable: true
                  routed_leg_count:
                    type: integer
                  routed_priced_leg_count:
                    type: integer
                  routed_pricing_coverage:
                    type: number
                    nullable: true
                    description: Fraction of actual DEX legs included in routed USD
                  round_trip_count:
                    type: integer
                  round_trip_gross_surplus_usd:
                    type: number
                    nullable: true
                    description: Successful on-chain round-trip surplus before transaction fees
                  round_trip_by_token:
                    type: array
                    items:
                      type: object
                      properties:
                        base_token:
                          type: string
                        tx_count:
                          type: integer
                        amount_in:
                          type: string
                        gross_surplus:
                          type: string
                        gross_surplus_usd:
                          type: number
                          nullable: true
                  round_trip_by_bridge:
                    type: array
                    description: Successful round trips grouped by intermediary bridge token
                    items:
                      type: object
                      properties:
                        bridge_token:
                          type: string
                        tx_count:
                          type: integer
                        amount_in:
                          type: string
                        gross_surplus:
                          type: string
                  by_token:
                    type: array
                    description: Entry notional grouped by invocation token_in
                    items:
                      type: object
                      properties:
                        token:
                          type: string
                        amount_in:
                          type: string
                          description: Entry notional in this token
                        routed_volume:
                          type: string
                          description: Actual input processed by DEX legs in this token
                        routed_leg_count:
                          type: integer
                  routed_by_token:
                    type: array
                    description: Actual DEX leg inputs grouped by each leg's token_in
                    items:
                      type: object
                      properties:
                        token:
                          type: string
                        routed_volume:
                          type: string
                        routed_leg_count:
                          type: integer
                  split_swap_count:
                    type: integer
                  success_count:
                    type: integer
                  failed_count:
                    type: integer
        error:
          type: string
...
    ArbitrageStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            granularity:
              type: string
              enum: [hour, day, week, month]
            start:
              type: integer
              format: int64
            end:
              type: integer
              format: int64
            buckets:
              type: array
              items:
                $ref: '#/components/schemas/ArbitrageStatsBucket'
        error:
          type: string
    ArbitrageStatsBucket:
      type: object
      required: [start, label, tx_count, success_count, failed_count, xlm_tx_count, usdc_tx_count, xlm_surplus, usdc_surplus]
      properties:
        start:
          type: integer
          format: int64
          description: Bucket start as a Unix timestamp in seconds.
        label:
          type: string
        tx_count:
          type: integer
          format: int64
        success_count:
          type: integer
          format: int64
        failed_count:
          type: integer
          format: int64
        xlm_tx_count:
          type: integer
          format: int64
        usdc_tx_count:
          type: integer
          format: int64
        xlm_surplus:
          type: string
          description: Successful XLM surplus in stroops.
        usdc_surplus:
          type: string
          description: Successful USDC surplus in smallest token units.
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional partner key for higher rate limits
security:
  - ApiKeyAuth: []